pyproject.toml
[tool.poetry]
name = "kenzie-commerce-api"
version = "0.1.0"
description = "projeto em equipe - m5"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "kenzie_commerce_api"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-spectacular = "^0.26.2"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipdb = "^0.13.13"
pycodestyle = "^2.10.0"
ipython = "^8.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
